refactor: Entreprise and Materiel class
